skill_analytics/requirements.txt
networkx
matplotlib
pydot
graphviz
pandas
python-dotenv == 1.0.1